+++ packages/smart_flutter_utils/test/utils/string_test.dart
@@ -5,32 +5,88 @@ import 'package:smart_flutter_utils/smart_flutter_utils.dart';
 void main() {
   group("StringConcatenation extension", () {
     const text = "hello";
+    const textBlank = " ";
+    const affix = "*";
 
-    test("String.pre()", () {
-      expect(text.pre("*"), "*hello");
-      expect(text.pre("*", doFor: 2), "**hello");
-      expect(text.pre("*", doFor: 3), "***hello");
-      expect(text.pre("*", between: " "), "* hello");
-      expect(text.pre(null), "hello");
-      expect(text.pre("*", doIf: false), "hello");
+    test("String.prefix()", () {
+      expect(text.prefix(affix), "*hello");
+      expect(text.prefix(affix, doFor: 2), "**hello");
+      expect(text.prefix(affix, doFor: 3), "***hello");
+
+      expect(text.prefix(affix, infix: " "), "* hello");
+      expect(text.prefix(affix, infix: " ", doFor: 2), "* * hello");
+
+      expect(
+        text.prefix(affix, infix: " ", suffix: "+"),
+        "* hello+",
+      );
+      expect(
+        text.prefix(affix, infix: " ", suffix: "+", doFor: 2),
+        "* * hello++",
+      );
+
+      expect(text.prefix(affix, doIf: false), text);
+      expect(text.prefix(null), text);
+      expect(text.prefix(" "), text);
+      expect(textBlank.prefix(affix), textBlank);
+      expect(textBlank.prefix(affix, notBlank: false), "* ");
     });
 
-    test("String.post()", () {
-      expect(text.post("*"), "hello*");
-      expect(text.post("*", doFor: 2), "hello**");
-      expect(text.post("*", doFor: 3), "hello***");
-      expect(text.post("*", between: " "), "hello *");
-      expect(text.post(null), "hello");
-      expect(text.post("*", doIf: false), "hello");
+    test("String.suffix()", () {
+      expect(text.suffix(affix), "hello*");
+      expect(text.suffix(affix, doFor: 2), "hello**");
+      expect(text.suffix(affix, doFor: 3), "hello***");
+
+      expect(text.suffix(affix, infix: " "), "hello *");
+      expect(text.suffix(affix, infix: " ", doFor: 2), "hello * *");
+
+      expect(
+        text.suffix(affix, infix: " ", prefix: "+"),
+        "+hello *",
+      );
+      expect(
+        text.suffix(affix, infix: " ", prefix: "+", doFor: 2),
+        "++hello * *",
+      );
+
+      expect(text.suffix(affix, doIf: false), text);
+      expect(text.suffix(null), text);
+      expect(text.suffix(" "), text);
+      expect(textBlank.suffix(affix), textBlank);
+      expect(textBlank.suffix(affix, notBlank: false), " *");
     });
 
-    test("String.surround()", () {
-      expect(text.surround("*"), "*hello*");
-      expect(text.surround("*", doFor: 2), "**hello**");
-      expect(text.surround("*", doFor: 3), "***hello***");
-      expect(text.surround("*", between: " "), "* hello *");
-      expect(text.surround(null), "hello");
-      expect(text.surround("*", doIf: false), "hello");
+    test("String.confix()", () {
+      expect(text.confix(affix), "*hello*");
+      expect(text.confix(affix, doFor: 2), "**hello**");
+      expect(text.confix(affix, doFor: 3), "***hello***");
+
+      expect(text.confix(affix, infix: " "), "* hello *");
+      expect(text.confix(affix, infix: " ", doFor: 2), "* * hello * *");
+
+      expect(
+        text.confix(affix, infix: " ", prefix: "+"),
+        "+* hello *",
+      );
+      expect(
+        text.confix(affix, infix: " ", prefix: "+", doFor: 2),
+        "++* * hello * *",
+      );
+
+      expect(
+        text.confix(affix, infix: " ", prefix: "+", suffix: "-"),
+        "+* hello *-",
+      );
+      expect(
+        text.confix(affix, infix: " ", prefix: "+", suffix: "-", doFor: 2),
+        "++* * hello * *--",
+      );
+
+      expect(text.confix(affix, doIf: false), text);
+      expect(text.confix(null), text);
+      expect(text.confix(" "), text);
+      expect(textBlank.confix(affix), textBlank);
+      expect(textBlank.confix(affix, notBlank: false), "* *");
     });
   });