+++ packages/smart_flutter_utils/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.0.4]
+
+* collection package dependency fixed.
+
+## [0.0.3]
+
+* Improved String concatenation and renamed as `prefix`, `suffix` and `confix`.
+* Added `mapToList` method
+
 ## [0.0.2]
 
 * Dart Format.